fix: prevent double anchoring (#266)
Fixes a bug caused by a bug in https://github.com/leanprover/lean4/pull/6730 that causes `infoFormatToHtml` to add nested anchor tags, with many of the outer links being unwanted. This PR modifies `infoFormatToHtml` to avoid wrapping HTML in anchor tags if the HTML already contains one.

This is an improvement on its own, but we will also address the root cause in a Lean PR.